Restructered files of location endpoints
