+++ server.js
@@ -38,7 +38,7 @@ var refCacheMiddleware = require('./server/utils/middleware/refCache');
 var helmet = require('helmet');
 var xssClean = require('xss-clean');
 var sanitizer = require('express-sanitizer');
-var locations = require('./server/routes/locations');
+var locations = require('./server/routes/locations/index');
 var postcodes = require('./server/routes/postcodes');
 var services = require('./server/routes/services');
 var registration = require('./server/routes/registration');
